extract btn. split styles
