+++ src/button.css
@@ -1,30 +1,3 @@
-*,
-:after,
-:before {
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-  padding: 0;
-  font-family: Inter, ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji",
-    "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
-}
-
-.root {
-  width: 100vw;
-  min-height: 100vh;
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-td {
-  text-align: center;
-  padding: 8px;
-}
-
 button {
   /* generic reset */
   font-family: inherit;
@@ -40,24 +13,23 @@ button {
   /* btn specific reset */
   background-color: transparent;
   background-image: none;
+  border: none;
 
   /* btn styles */
   cursor: pointer;
 }
 
 .btn {
-  border: none; /* move to reset? */
   display: inline-flex;
-  padding: 10px;
   justify-content: center;
   align-items: center;
 
   /* typography */
-  font-family: Inter;
+  /* font-family: Inter;
   font-size: 16px;
   font-style: normal;
   font-weight: 400;
-  line-height: normal;
+  line-height: normal; */
 }
 
 .btn:disabled {
@@ -148,4 +120,8 @@ button {
  ******************/
 .btn--small {
   padding: 4px;
+}
+
+.btn--base {
+  padding: 10px;
 }
